improvement: move format icns name to utils folder
