+++ packages/web/utils/string.ts
@@ -19,3 +19,20 @@ export function getShortAddress(
     )
   );
 }
+
+export const formatICNSName = (name?: string, maxLength = 28) => {
+  if (!name) return undefined;
+  if (name.length <= maxLength) return name;
+
+  const nameParts = name.split(".");
+  const userName = nameParts[0];
+  const chain = nameParts[1];
+
+  return (
+    userName.substring(0, 10) +
+    "..." +
+    userName.substring(userName.length - 5, userName.length) +
+    "." +
+    chain
+  );
+};
